fix: some more refactoring
